캐릭터데이터 및 스크립트 수정 & 골렘 모델링 및 애니메이션 추가
1. CharacterData 스크립트 1차 정리
2. 정리 후 에러 해결을 위한 CharacterManager, CombatManager, BaseMonsterAI, PlayerController 스크립트 일부 수정
3. 골렘 몬스터 모델 추가 및 애니메이션 추가
4. WeaponCollider 스크립트 딜레이 늘림 0.1 -> 0.5

+++ Assets/DsProject/Scripts/Managers/CombatManager.cs
@@ -64,14 +64,14 @@ public class CombatManager : BaseManager<CombatManager>
     private int CalculateDamage(CharacterData attacker, CharacterData defender)
     {
         // 공격자의 물리 공격력, 방어자의 물리 방어력 계산 로직
-        int damage = attacker.physicalDamage - defender.physicalDefense;
+        float damage = attacker.physicalDamage - defender.physicalDefense;
 
         // 크리티컬 여부 체크
         bool isCritical = Random.value < attacker.criticalChance;
         damage = attacker.CalculateDamage(isCritical);
         
         // 데미지는 최소 0 이상
-        return Mathf.Max(0, damage);
+        return (int)Mathf.Max(0, damage);
     }
 
     // 몬스터 사망 처리
